챌린지 인증 리턴값 형식 수정

+++ gbt/src/main/java/ksafinalproject/gbt/smoking/controller/SmokingController.java
@@ -45,12 +45,12 @@ public class SmokingController {
     }
 
     @PostMapping("/challenge-attend")
-    public Long smokingChallengeAttendSave(@RequestBody ISmoking iSmoking) {
+    public int smokingChallengeAttendSave(@RequestBody ISmoking iSmoking) {
         try {
             return smokingService.saveChallengeAttendSmoking(iSmoking);
         } catch (Exception e) {
             log.error("Error : {}", e.getMessage());
-            return -1L;
+            return -1;
         }
     }
